Moves FeedImagePresenter and collaborators to production.
